% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtest_results.R
\name{as_dbtest_results}
\alias{as_dbtest_results}
\alias{as.dbtest_results}
\title{Coerce object to a dbtest_results object}
\usage{
as_dbtest_results(x, ...)

as.dbtest_results(x, ...)
}
\arguments{
\item{x}{A list.  Should have a "results" object of class
`testthat_results` and a "connection" object of class `character`}

\item{...}{Other arguments passed on to individual methods}
}
\description{
`as_dbtest_results` will convert a list to a dbtest_results
object, provided that it meets the necessary requirements
}
