% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_test_that.R
\name{db_test_con}
\alias{db_test_con}
\title{Create a connection to a test database for testing database transactions.}
\usage{
db_test_con(...)
}
\arguments{
\item{...}{list. Additional arguments to pass to /code{db_connection}}
}
\value{
a database connection suitable for testing.
}
\description{
Create a connection to a test database for testing database transactions.
}

