% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_test_that.R
\name{db_test_that}
\alias{db_test_that}
\title{Run a single testthat test in a self-contained database transaction.}
\usage{
db_test_that(description, test)
}
\arguments{
\item{description}{character. The description of the test.}

\item{test}{expression. The test to execute.}
}
\description{
This will contain your database actions to a single test, by deleting all the
tables in the test db before you begin.  This keeps tests from interfering
with each other.
}

