% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtest.R
\docType{package}
\name{dbtest-package}
\alias{dbtest}
\alias{dbtest-package}
\title{dbtest: Test automation for dbplyr SQL translation}
\description{
The package uses testthat to automate tests of dbplyr SQL translation
by running the tests against live database connections.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/dbtest}
  \item Report bugs at \url{https://github.com/rstudio/dbtest/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgar@rstudio.com}

Authors:
\itemize{
  \item Cole Arendt \email{cole@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
