% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db_test_that.R
\name{drop_all_test_tables}
\alias{drop_all_test_tables}
\title{Drop all the tables in the test database.}
\usage{
drop_all_test_tables(test_con)
}
\arguments{
\item{test_con}{DBIConnection. A connection to the test database.}
}
\description{
Drop all the tables in the test database.
}

