% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_sql_exists}
\alias{expect_sql_exists}
\title{Expect that a certain SQL statement will have a result.}
\usage{
expect_sql_exists(statement)
}
\arguments{
\item{statement}{character. The SQL statement to run.}
}
\description{
Expect that a certain SQL statement will have a result.
}

