% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_sql_is}
\alias{expect_sql_is}
\title{Expect that a certain SQL statement will evaluate to a result.}
\usage{
expect_sql_is(statement, expected)
}
\arguments{
\item{statement}{character. The SQL statement to run.}

\item{expected}{ANY. The expected result.}
}
\description{
Expect that a certain SQL statement will evaluate to a result.
}
\examples{
\dontrun{expect_sql_is("SELECT id FROM flights LIMIT 1", 1)}
}

