% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_table}
\alias{expect_table}
\title{Expect that the table exists in the test database.}
\usage{
expect_table(table_name)
}
\arguments{
\item{table_name}{character. The name of the table to test for.}
}
\description{
Expect that the table exists in the test database.
}
\examples{
\dontrun{expect_table("flights")}
}

