% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_table_has}
\alias{expect_table_has}
\title{Expect that the table has a certain property.}
\usage{
expect_table_has(test, table)
}
\arguments{
\item{test}{ANY. The test to run.}

\item{table}{character. The name of the table to check.}
}
\description{
Expect that the table has a certain property.
}
\examples{
\dontrun{
  expect_table_has(column("id"), table = "flights")
  expect_table_has(count("id") > 0, table = "flights")
}
}

