% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{get_dbtest_detail}
\alias{get_dbtest_detail}
\title{Get dbtest Detail}
\usage{
get_dbtest_detail(.obj, db = NULL, file = NULL, context = NULL,
  verb = NULL)
}
\arguments{
\item{.obj}{The dbtest_results object (or a list of such objects) to get detail for}

\item{db}{optional The database label to filter by}

\item{file}{optional The test file to filter by}

\item{context}{optional The test context to filter by}

\item{verb}{optional The verb to filter by}
}
\value{
A tibble with test details (and stack traces)
}
\description{
Retrieve test details from a list of dbtest_results objects.
Returns a much more natural object to work with than nested lists
and enables more easily surfacing the _reasons_ that tests
failed.
}
