% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtest_results.R
\name{is.dbtest_results}
\alias{is.dbtest_results}
\alias{is_dbtest_results}
\title{Test if the object is a dbtest_results object}
\usage{
is.dbtest_results(x)

is_dbtest_results(x)
}
\arguments{
\item{x}{An object}
}
\description{
This function returns FALSE if the object is not a
dbtest_results object, and TRUE if it is.
}
