% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycflights13_sql.R
\name{nycflights13_sql}
\alias{nycflights13_sql}
\title{Create an in-memory SQLite database for testing}
\usage{
nycflights13_sql(con, schema = "public", ...)
}
\arguments{
\item{con}{an SQL connection (i.e a PostgreSQL connection)}

\item{schema}{schema to write the tables ("public" by default)}

\item{...}{additional parameters to connect to a database}
}
\value{
writes to remote database
}
\description{
Create an in-memory SQLite database for testing
}
\examples{
\dontrun{
# psql_con is a PostgreSQL connnection from .Rprofile
nycflights13_sql(psql_con, schema = "nycflights13")
}
}
