% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{plot_tests}
\alias{plot_tests}
\alias{plot_summary}
\title{Plot Tests}
\usage{
plot_tests(results)

plot_summary(results)
}
\arguments{
\item{results}{Output from `test_single_database` or `test_database`}
}
\value{
list of ggplot2 objects / graphs
}
\description{
Plot the output from `test_single_database` or `test_database` as a ggplot2 object
for easy visualization of test success or failure across databases.
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
tbl_data <- dplyr::copy_to(con, testdata)
res <- test_database(tbl_data, pkg_test("simple-tests.yml"))
plot_tests(res)
DBI::dbDisconnect(con)

}
