% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{print_interactive}
\alias{print_interactive}
\alias{print_interactive.list}
\alias{print_interactive.default}
\title{Print Interactively}
\usage{
print_interactive(.obj, .interactive = interactive())

\method{print_interactive}{list}(.obj, .interactive = interactive())

\method{print_interactive}{default}(.obj, .interactive = interactive())
}
\arguments{
\item{.obj}{The object which needs to be printed interactively}

\item{.interactive}{Whether to prompt for interactivity}
}
\value{
List of printed objects
}
\description{
Prints a list object by executing print on each element.
If in an interactive session, prompts to continue after
each print.
}
