% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_yml}
\alias{read_yml}
\title{Reads a yaml file from a specified path.}
\usage{
read_yml(path)
}
\arguments{
\item{path}{character. The file name of the yaml file.}
}
\value{
a nested list as specified by the yaml file.
}
\description{
Reads a yaml file from a specified path.
}

