% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_database}
\alias{test_database}
\alias{test_databases}
\title{Test Database}
\usage{
test_database(datasource = NULL, tests = pkg_test(), return_list = TRUE,
  skip = NULL)

test_databases(datasource = NULL, tests = pkg_test())
}
\arguments{
\item{datasource}{optional Defaults to using a SQLite database.  Pass "dsn" to use
all DSNs available on the system.  Use "config" or a path to a "config.yml" file to
use connection parameters in a YAML file.  Connection parameters will be passed to
`dbConnect` as-is}

\item{tests}{optional  A character vector of yaml tests to execute.
References `dbtest` test suite by default}

\item{return_list}{optional Whether to return a list of `dbtest_results` objects. Defaults
to TRUE.  Provide FALSE if you desire a single database test to return a `dbtest_results`
object directly.}
}
\value{
Returns a list of lists containing the respective datasource labels and testthat output
}
\description{
A test executor that iterates over multiple datasources
and executes the testing suite on each.  Output is organized in such a way as to
give nice, consolidated results.
}
\examples{
# test all dsns with default test suite -----------------------
\dontrun{
test_database(datasource = "dsn")
}
# test sqlite with custom suite -------------------------
\dontrun{
test_database(tests = "./path/to/my.yml")
}
# test connection yaml file with default test suite -----------
\dontrun{
test_database(datasource = "./path/to/conn.yml")
}

}
