% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_databases}
\alias{test_databases}
\title{Test Databases}
\usage{
test_databases(datasources = NULL, tests = pkg_test())
}
\arguments{
\item{datasources}{optional Defaults to using a SQLite database.  Pass "dsn" to use
all DSNs available on the system.  Use "config" or a path to a "config.yml" file to
use connection parameters in a YAML file.  Connection parameters will be passed to
`dbConnect` as-is}

\item{tests}{optional The tests to execute.  References `dbtest` test suite by default}
}
\value{
Returns a list of lists containing the respective datasource labels and testthat output
}
\description{
A wrapper around `test_sigle_database` that iterates over multiple datasources
and executes the testing suite on each.  Output is organized in such a way as to
give nice, consolidated results.
}
\examples{
# test all dsns with dbtest suite -----------------------
\dontrun{
test_databases(datasources = "dsn")
}
# test sqlite with custom suite -------------------------
\dontrun{
test_databases(tests = "./path/to/my.yml")
}
# test connection yaml file with dbtest suite -----------
\dontrun{
test_databases(datasources = "./path/to/conn.yml")
}

}
\seealso{
test_single_database
}
