% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_single_database}
\alias{test_single_database}
\title{Test Single Database}
\usage{
test_single_database(datasource, label = NULL, tests = default_test_path())
}
\arguments{
\item{datasource}{The datasource to test against.  Either a DBI connection or a tbl_sql}

\item{label}{optional The label to give the test.  If not provided, one will be generated}

\item{tests}{optional The tests to execute.  References `dbtest` test suite by default}
}
\value{
A list object with the label and testthat results
}
\description{
Run a single datasource through the testing suite.  Typically, this
object would be a connection or a `tbl_sql`
}
\examples{

con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
res <- test_single_database(con, "simple-tests.yml")
DBI::dbDisconnect(con)

}
\seealso{
test_databases
}
