% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Test Data}
\format{A data frame with 10 rows and 9 columns:
\describe{
  \item{fld_factor}{A field with values in "a","b","c"}
  \item{fld_datetime}{A field with datetime values}
  \item{fld_date}{A field with date values}
  \item{fld_time}{A field with time values}
  \item{fld_binary}{A field with 1s or 0s}
  \item{fld_integer}{A field with integer values}
  \item{fld_double}{A field with double precision values}
  \item{fld_character}{A field with character values}
  \item{fld_logical}{A field with logical values}
}}
\usage{
testdata
}
\description{
A simple dataset that presents varied data types to facilitate
testing
}
\keyword{datasets}
