% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testdata_large}
\alias{testdata_large}
\title{Test Data Large}
\format{A data frame with 150 rows and 5 columns:
\describe{
  \item{fld_double}{A field with double precision values}
  \item{fld_integer}{A field with integer values}
  \item{fld_character}{A field with character values}
  \item{fld_binary}{A field with binary (1 or 0) values}
  \item{fld_factor}{A field with factor values}
}}
\usage{
testdata_large
}
\description{
A test dataset with slightly more data.  At present this is just
a variation of good ol' iris with renamed fields
}
\keyword{datasets}
