% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_tests_path}
\alias{default_tests_path}
\alias{default_config_path}
\title{Utility Functions}
\usage{
default_tests_path(file = "tests.yml")

default_config_path(file = "config.yml")
}
\arguments{
\item{file}{The file to reference at the default location}
}
\description{
Utility functions useful for package default behavior
}
