% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbtest.R
\name{with_test_db}
\alias{with_test_db}
\title{Run a block of code where all DBI connections are mocked with the test connection.}
\usage{
with_test_db(expr)
}
\arguments{
\item{expr}{expression. A block of R code to run.}
}
\description{
Run a block of code where all DBI connections are mocked with the test connection.
}

