% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_test}
\alias{write_test}
\title{Write Tests}
\usage{
write_test(file, header, expr, overwrite = FALSE, comparison = " > 10")
}
\arguments{
\item{file}{The file the tests should be written to}

\item{header}{The header to name the test section.  Preferably short}

\item{expr}{The expression to use for writing tests}

\item{overwrite}{optional Whether to overwrite the existing file or
append.  Defaults to FALSE}

\item{comparison}{optional The type of comparison to use for filter
when creating the test.  Currently ignored}
}
\description{
A utility to make writing tests a bit easier
and more reproducible
}
