% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dbtest.R
\name{db_test_con}
\alias{db_test_con}
\title{Create a connection to a test database for testing database transactions.}
\usage{
db_test_con()
}
\value{
a database connection suitable for testing.
}
\description{
Create a connection to a test database for testing database transactions.
}

