% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{test_single_database}
\alias{test_single_database}
\title{Test Single Database}
\usage{
test_single_database(datasource, tests = pkg_test(), label = NULL)
}
\arguments{
\item{datasource}{The datasource to test against.  Either a DBI connection or a tbl_sql}

\item{tests}{optional A character vector of yaml tests to execute.
References `dbtest` test suite by default}

\item{label}{optional The label to give the test.  If not provided, one will be generated}
}
\value{
A list object with the label and testthat results
}
\description{
Run a single datasource through the testing suite.  Typically, this
object would be a connection or a `tbl_sql`
}
\examples{

\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
res <- test_single_database(con, pkg_test("simple-tests.yml"))
DBI::dbDisconnect(con)
}

}
\seealso{
test_database
}
