% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_test_path}
\alias{default_test_path}
\alias{all_tests}
\alias{default_config_path}
\title{Utility Functions}
\usage{
default_test_path(file = "simple-tests.yml")

all_tests(dir = system.file("extdata", "tests", package = "dbtest"))

default_config_path(file = "config.yml")
}
\arguments{
\item{file}{optional The file to reference at the default location}

\item{dir}{optional The directory for which to show all_tests}
}
\description{
Utility functions useful for package default behavior.
`default_test_path` returns the path to tests from the `dbtest` package.
`default_config_path` returns the path to connection config from the `dbtest`
package.  `all_tests` returns the path to all .yml or .yaml files in a given
directory.  The default is to return all tests in the `dbtest` package
}
