% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_test}
\alias{pkg_test}
\alias{all_tests}
\alias{pkg_config}
\alias{pkg_skip}
\title{Utility Functions}
\usage{
pkg_test(file = "simple-tests.yml")

all_tests(dir = system.file("extdata", "tests", package = "dbtest"))

pkg_config(file = "config.yml")

pkg_skip(file = "skip-ex.yml")
}
\arguments{
\item{file}{optional The file to reference at the default location}

\item{dir}{optional The directory for which to show all_tests}
}
\description{
Utility functions useful for package default behavior.
`pkg_test` returns the path to tests from the `dbtest` package.
`pkg_config` returns the path to connection config from the `dbtest`
package.  `all_tests` returns the path to all .yml or .yaml files in a given
directory.  The default is to return all tests in the `dbtest` package
}
