% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_test}
\alias{write_test}
\title{Write Tests}
\usage{
write_test(file, header, expr, overwrite = FALSE, comparison = .x >
  sample(1:10, 1))
}
\arguments{
\item{file}{The file the tests should be written to}

\item{header}{The header to name the test section.  Preferably short}

\item{expression}{The expression to use for writing tests}
}
\description{
A utility to make writing tests a bit easier
and more reproducible
}
